+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Features
 Bug Fixes
 ---------
 
+* Fix db test bug with month addition to properly handle December. Issue #261.
 * Relax endpoint TLS cert requirement for https scheme. Issue #249.
 
 WebPush